third_party/blink/web_tests/svg/dynamic-updates/SVGFilterElement-svgdom-primitiveUnits-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'primitiveUnits' property of the SVGFilterElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS filterElement.primitiveUnits.baseVal is SVGUnitTypes.SVG_UNIT_TYPE_USERSPACEONUSE
PASS filterElement.primitiveUnits.baseVal is SVGUnitTypes.SVG_UNIT_TYPE_OBJECTBOUNDINGBOX
PASS successfullyParsed is true

TEST COMPLETE
